refactor : refactor for DataBinding on PlacesAdapter
